--- src/main/java/ladderstep4/ladder/domain/LadderString.java
@@ -1,5 +1,0 @@
-package ladderstep4.ladder.domain;
-
-public interface LadderString {
-    String value();
-}
