--- src/main/java/ladderstep4/ladder/strategy/MakeRungStrategy.java
@@ -1,8 +1,0 @@
-package ladderstep4.ladder.strategy;
-
-@FunctionalInterface
-public interface MakeRungStrategy {
-
-    boolean test();
-
-}
