--- src/main/java/ladderstep4/ladder/strategy/RandomRungSupplier.java
@@ -1,14 +1,0 @@
-package ladderstep4.ladder.strategy;
-
-import java.util.Random;
-
-public class RandomRungSupplier implements MakeRungStrategy {
-
-    private static final Random RANDOM = new Random();
-
-    @Override
-    public boolean test() {
-        return RANDOM.nextBoolean();
-    }
-
-}
